+++ test/org/moeaframework/core/indicator/PISAHypervolumeTest.java
@@ -25,6 +25,7 @@ import org.moeaframework.core.Problem;
 import org.moeaframework.core.Settings;
 import org.moeaframework.core.Solution;
 import org.moeaframework.core.spi.ProblemFactory;
+import org.moeaframework.util.PropertyScope;
 
 /**
  * Tests the {@link PISAHypervolume} class. Due to performance, these tests only go up to 4 dimensions.
@@ -182,17 +183,13 @@ public class PISAHypervolumeTest extends IndicatorTest {
 	
 	@Test
 	public void testExplicitBounds2_Properties() {
-		Settings.PROPERTIES.setDouble("org.moeaframework.core.indicator.hypervolume_idealpt.DTLZ2", 0.0);
-		Settings.PROPERTIES.setDouble("org.moeaframework.core.indicator.hypervolume_refpt.DTLZ2", 2.0);
-		
-		Problem problem = ProblemFactory.getInstance().getProblem("DTLZ2_2");
-		
-		PISAHypervolume hypervolume = new PISAHypervolume(problem, new NondominatedPopulation());
-		
-		test2(hypervolume);
-		
-		Settings.PROPERTIES.remove("org.moeaframework.core.indicator.hypervolume_idealpt.DTLZ2");
-		Settings.PROPERTIES.remove("org.moeaframework.core.indicator.hypervolume_refpt.DTLZ2");
+		try (PropertyScope scope = Settings.createScope()
+				.with("org.moeaframework.core.indicator.hypervolume_idealpt.DTLZ2", 0.0)
+				.with("org.moeaframework.core.indicator.hypervolume_refpt.DTLZ2", 2.0)) {
+			Problem problem = ProblemFactory.getInstance().getProblem("DTLZ2_2");
+			PISAHypervolume hypervolume = new PISAHypervolume(problem, new NondominatedPopulation());
+			test2(hypervolume);
+		}
 	}
 
 }